Fixed all the trivial warnings. Down to about 30 or so.
git-svn-id: https://svn.java.net/svn/jna~svn/trunk@1076 2f8a963e-d2e4-e7d0-97bf-ccb7fcea9d80

+++ jnalib/test/com/sun/jna/StructureByValueTest.java
@@ -91,14 +91,14 @@ public class StructureByValueTest extends TestCase {
     }
     public void testStructureArgByValue64() {
         ByValue64 data = new ByValue64();
-        final long DATA = (long)MAGIC;
+        final long DATA = MAGIC;
         data.data = DATA;
         assertEquals("Failed to pass 64-bit struct by value",
                      DATA, lib.testStructureByValueArgument64(data));
     }
     public void testStructureArgByValue128() {
         ByValue128 data = new ByValue128();
-        final long DATA = (long)MAGIC;
+        final long DATA = MAGIC;
         data.data = DATA;
         data.data1 = DATA;
         assertEquals("Failed to pass 128-bit struct by value",
